requirements.txt
pandas
flask
numpy
scikit-learn
pymongo
pyyaml
-e .